cloud-run/finance-operations/requirements.txt
Flask==3.0.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
gunicorn==21.2.0
